// types/employee.ts
export interface Employee {
    id:          number;
    name:        string;
    place_id:    number;
    place_name:  string;
    birth_place: string;
    birth_date:  Date;
    address:     string;
    phone:       string;
    photo:       string;
    email:       string;
    status:      string;
    created_on:  Date;
    updated_on:  null;
    deleted:     number;
}